Create PageRequestHandler, which extends standard RequestHandler.
This change makes class RequestHandler a simple handler that does not
know about pages or HTTP responses. The only method that must be
implemented by RequestHandler sub-classes is handle(). A handler can
implement, for example, a simple action.

The new class PageRequestHandler can be used for handling page
request, which means requests that expect an HTTP response
containing an HTML document.